v3.0

+++ starter/bin/deploy_oke.sh
@@ -109,11 +109,13 @@ kubectl apply -f $TARGET_OKE/tf_env_configmap.yaml
 # APP
 for APP_NAME in `app_name_list`; do
   APP_YAML="k8s_${APP_NAME}.yaml"
-  cp src/app/${APP_YAML} $TARGET_OKE/.
-  # If present, replace the ORDS URL
-  ORDS_HOST=`basename $(dirname $ORDS_URL)`
-  file_replace_variables $TARGET_OKE/${APP_YAML}
-  kubectl apply -f $TARGET_OKE/${APP_YAML}
+  if [ -f $APP_YAML ]; then
+    cp src/app/${APP_YAML} $TARGET_OKE/.
+    # If present, replace the ORDS URL
+    ORDS_HOST=`basename $(dirname $ORDS_URL)`
+    file_replace_variables $TARGET_OKE/${APP_YAML}
+    kubectl apply -f $TARGET_OKE/${APP_YAML}
+  fi
 done
 
 # UI
